Moves detection upgrade & Checkmate detection
Implemented a check on the king to prevent the king on moving onto an attacked square. Also added CheckMate detection.

+++ src/main/GamePanel.java
@@ -161,7 +161,7 @@ public class GamePanel extends JPanel implements Runnable, GameObserver {
     }
 
     public void run() {
-        double drawInterval = 1000000000 / fps;
+        double drawInterval = (double) 1000000000 / fps;
         double delta = 0;
         long lastTime = System.nanoTime();
         long currentTime;
